refactor: application에서 비즈니스 요구사항인 예외를 처리해주도록 리팩토링

+++ src/main/java/wooteco/subway/line/service/LineService.java
@@ -21,18 +21,30 @@ public class LineService {
     }
 
     public Line save(final Line line) {
+        if (lineRepository.doesNameExist(line)) {
+            throw new DuplicateLineNameException();
+        }
         return lineRepository.save(line);
     }
 
     public Line getLine(final Long id) {
+        if (lineRepository.doesIdNotExist(id)) {
+            throw new NoSuchLineException();
+        }
         return lineRepository.getLine(id);
     }
 
     public void updateLine(final Line line) {
+        if (lineRepository.doesIdNotExist(line)) {
+            throw new NoSuchLineException();
+        }
         lineRepository.update(line);
     }
 
     public void deleteById(final Long id) {
+        if (lineRepository.doesIdNotExist(id)) {
+            throw new NoSuchLineException();
+        }
         lineRepository.deleteById(id);
     }
 }
